pyproject.toml
[tool.poetry]
name = "gadoneko"
version = "1.5.0"
description = "Just an overegnieered discord bot for students."
authors = ["Krystian Postek <krystian@postek.eu>"]

[tool.poetry.dependencies]
python = ">=3.10,<3.11"
# Bot related dependencies
hikari = {extras = ["speedups"], version = "^2.0.0-alpha.104"}
hikari-lightbulb = ">=2.0.0"
lightbulb-ext-neon = {git = "https://github.com/neonjonn/lightbulb-ext-neon.git"}
# Cron related dependencies
aiocron = "*"
# Web related dependencies
starlette = "*"
python-multipart = "*"
jinja2 = "*"
uvicorn = "*"
# Common dependencies
mongoengine = ">=0.23.1"
pymongo = "<4.0"
python-dotenv = "*"
google-api-python-client = "*"
sentry-sdk = "*"
aiohttp = "<4.0.0"
pyyaml = "*"
millify = "*"
itsdangerous = "^2.0.1"
starlette-discord = "^0.2.0"
markdownify = "^0.10.1"

[tool.poetry.dev-dependencies]
black = "*"
pylint = "*"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
